散點圖串資料庫資料

+++ app/src/main/java/com/example/purpulse/Note.java
@@ -5,5 +5,5 @@ import java.util.ArrayList;
 public class Note {
     public static String account,name,mail,birthday,gender,height,weight;
     public static Double sdNN,LFn,HFn;
-    public static ArrayList<Double> RRi;
+    public static ArrayList<Float> RRi;
 }
